Full Parameterized Modules completed
The design of the module Register File has been changed to a Structural
design using different modules, these modules has been design using
parameters to indicate the lenght of the words that will be stored, and
the number of bits to use as directions. Also the register $ZERO has
been defined.

Andres Rivera Marquez.